unify WASM and native diagnostic implementations into shared diagnostics_core module
